wire up more templates
