+++ components/quiz/answerWrapper.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function AnswerWrapper({children}) {
+  return <div>{children}</div>
+}
