--- components/quiz/components/index.js
@@ -1,6 +1,0 @@
-import MultipleChoice from './multipleChoice'
-import Essay from './essay'
-import Theater from './theater'
-import Sketch from './sketch'
-
-export {Essay, MultipleChoice, Sketch, Theater}
