+++ components/quiz/explanation.js
@@ -1,0 +1,3 @@
+export default function Explanation({children}) {
+  return <div>{children}</div>
+}
