+++ components/quiz/questionWrapper.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function QuestionWrapper({children}) {
+  return <div>{children}</div>
+}
