+++ components/quiz/sketch.js
