+++ components/quiz/wrapper.js
@@ -1,0 +1,3 @@
+export default function Wrapper({children}) {
+  return <div>{children}</div>
+}
