+++ hooks/useEggheadQuizMachine.js
@@ -1,4 +1,3 @@
-import {useLocalStorage} from 'react-use'
 import {useMachine} from '@xstate/react'
 import {quizMachine} from 'machines/quizMachine'
 import {first, indexOf, find, get} from 'lodash'
@@ -18,21 +17,20 @@ export default function useEggheadQuizMachine(slug) {
     (currentQuestionIdx + 1 === questions.length
       ? get(first(questions), 'id')
       : get(questions[currentQuestionIdx + 1], 'id'))
+  const isCurrentQuestionAnswered = find(state.context.answers, {
+    id: currentQuestionId,
+  })
+  const isDisabled = state.matches('answering') || state.matches('answered')
 
   function handleContinue() {
     send('NEXT_QUESTION', {nextQuestionId: nextQuestionId})
   }
   function handleSubmit(values, actions) {
-    send('SUBMIT', {answer: {...values}, question: {...currentQuestion}})
-    // window.alert(JSON.stringify(values, null, 2))
-    // send('NEXT_QUESTION', {nextQuestionId: nextQuestionId})
-    // const now = Date.now()
-    // const date = new Date(now).toUTCString()
-    // const context = {} // {quizId: quiz.id, questionId: question.id, date}
+    const now = Date.now()
+    const date = new Date(now).toUTCString()
+    // const context = {quizId: quiz.id, questionId: question.id, date}
     // const response = {...values, question, context}
-    // console.log({response})
-    // window.alert(JSON.stringify(response, null, 2))
-    // markCompleted(question)
+    send('SUBMIT', {answer: {...values, ...currentQuestion}})
   }
 
   return {
@@ -42,5 +40,7 @@ export default function useEggheadQuizMachine(slug) {
     send,
     handleContinue,
     handleSubmit,
+    isDisabled,
+    isCurrentQuestionAnswered,
   }
 }
